add AddIngredient, refactor to use reducer in a custom hook for ingredients
