+++ src/Ingredients/AddIngredient.test.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import Ingredients from './Ingredients'
+import AddIngredient from "./AddIngredient";
 
 test("renders properly", () => {
-  const { container } = render(<Ingredients />);
+  const { container } = render(<AddIngredient handleSubmit={() => {}} />);
   expect(container).toMatchSnapshot();
 });
